GRPCValidService/requirements.txt
grpcio
grpcio-tools
opencv-python-headless
numpy
python-dotenv
